# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"properties": {
					"Id": {
						"type": "number",
						"format": "double"
					},
					"IsActive": {
						"type": "boolean"
					},
					"CreatedDate": {
						"type": "string"
					},
					"UpdatededDate": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Dob": {
						"type": "string"
					},
					"PhoneNumber": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Id",
					"IsActive",
					"CreatedDate",
					"UpdatededDate",
					"FirstName",
					"LastName",
					"Dob",
					"PhoneNumber",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"anyOf": [
						{
							"type": "string"
						},
						{
							"type": "number",
							"format": "double"
						},
						{
							"properties": {},
							"type": "object"
						}
					]
				}
			},
			"ICommonResponse_IUser-Array_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"body": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							},
							{
								"$ref": "#/components/schemas/IObject"
							}
						]
					},
					"message": {
						"type": "string"
					},
					"error": {
						"$ref": "#/components/schemas/IObject"
					}
				},
				"required": [
					"status",
					"body",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICommonResponse____": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"body": {
						"anyOf": [
							{
								"properties": {},
								"type": "object"
							},
							{
								"$ref": "#/components/schemas/IObject"
							}
						]
					},
					"message": {
						"type": "string"
					},
					"error": {
						"$ref": "#/components/schemas/IObject"
					}
				},
				"required": [
					"status",
					"body",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserRequestDto": {
				"properties": {
					"Id": {
						"type": "number",
						"format": "double"
					},
					"IsActive": {
						"type": "boolean"
					},
					"CreatedDate": {
						"type": "string"
					},
					"UpdatededDate": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Dob": {
						"type": "string"
					},
					"PhoneNumber": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"ConfirmPassword": {
						"type": "string"
					}
				},
				"required": [
					"Id",
					"IsActive",
					"CreatedDate",
					"UpdatededDate",
					"FirstName",
					"LastName",
					"Dob",
					"PhoneNumber",
					"Email",
					"Password",
					"ConfirmPassword"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nodetsapicustom",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/auth/users": {
			"get": {
				"operationId": "Users",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICommonResponse_IUser-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/auth/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICommonResponse____"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserRequestDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}